==> practico2/01_explorer_data.ipynb <==
### Conjunto de datos de predicción de riesgo de ataque cardíaco

El conjunto de datos de **clasificación de pacientes con riesgo de ataque cardíaco** contiene **8.763 registros** de pacientes de diferentes países con una columna respuesta que se centra en la presencia o ausencia de riesgo de ataque cardíaco. 
Incluye atributos como:

- **Patient ID** - Identificador único para cada paciente
- **Age** - Edad del paciente
- **Sex** - Género del paciente (Masculino/Femenino)
- **Cholesterol** - Niveles de colesterol del paciente
- **Blood Pressure** - Presión arterial del paciente (sistólica/diastólica)
- **Heart Rate** - Frecuencia cardíaca del paciente
- **Diabetes** - Si el paciente tiene diabetes (Sí/No)
- **Family History** - Historial familiar de problemas cardíacos (1: Sí, 0: No)
- **Smoking** - Estado de tabaquismo del paciente (1: Fumador, 0: No fumador)
- **Obesity** - Estado de obesidad del paciente (1: Obeso, 0: No obeso)
- **Alcohol Consumption** - Nivel de consumo de alcohol por el paciente
- **Exercise Hours Per Week** - Número de horas de ejercicio por semana
- **Diet** - Hábitos alimenticios del paciente (Saludable/Promedio/No saludable)
- **Previous Heart Problems** - Problemas cardíacos previos del paciente (1: Sí, 0: No)
- **Medication Use** - Uso de medicamentos por el paciente (1: Sí, 0: No)
- **Stress Level** - Nivel de estrés reportado por el paciente (1-10)
- **Sedentary Hours Per Day** - Horas de actividad sedentaria por día
- **Income** - Nivel de ingresos del paciente
- **BMI** - Índice de Masa Corporal (IMC) del paciente
- **Triglycerides** - Niveles de triglicéridos del paciente
- **Physical Activity Days Per Week** - Días de actividad física por semana
- **Sleep Hours Per Day** - Horas de sueño por día
- **Country** - País del paciente
- **Continent** - Continente donde reside el paciente
- **Hemisphere** - Hemisferio donde reside el paciente
- **Heart Attack Risk** - Presencia de riesgo de ataque cardíaco (1: Sí, 0: No)

import numpy as np
import pandas as pd

raw_data = pd.read_csv("heart_attack_prediction_dataset.csv")
raw_data.head(5)

raw_data.shape

raw_data.describe()

#### Verificar valores nulos

raw_data.info()

raw_data.isnull().sum()

#### Revisar filas duplicadas

raw_data.drop_duplicates(keep=False).shape

#### Separar columna de presión arterial

raw_data[['Systolic', 'Diastolic']] = raw_data['Blood Pressure'].str.split('/', expand=True)
raw_data[['Systolic', 'Diastolic']] = raw_data[['Systolic', 'Diastolic']].astype(int)
raw_data.drop('Blood Pressure', axis=1, inplace=True)

#### Eliminar la columna del ID

raw_data.drop('Patient ID', axis=1, inplace=True)

#### Balance de las clases de la respuesta

raw_data["Heart Attack Risk"].value_counts()

#### Undersampling para la clase mayoritaria y balancear el dataset

class_1 = raw_data[raw_data['Heart Attack Risk'] == 1]
class_0 = raw_data[raw_data['Heart Attack Risk'] == 0]

# Realizar undersampling en la clase con más datos (en este caso, la clase 0)
class_0_sampled = class_0.sample(n=len(class_1), random_state=42)

balanced_data = pd.concat([class_1, class_0_sampled])
balanced_data = balanced_data.sample(frac=1, random_state=42).reset_index(drop=True)

balanced_data

balanced_data["Heart Attack Risk"].value_counts()

balanced_data.to_csv('process_dataset/01_first_result_balanced.csv', index=False)

raw_data.to_csv('process_dataset/01_first_result.csv', index=False)
