==> practico2/03_evaluation_outliers.ipynb <==
import pandas as pd
from sklearn.ensemble import IsolationForest

#### Detección de valores atípicos

df_data = pd.read_csv('process_dataset/02_second_result.csv')

def detect_outliers_iqr(df):
    outliers_list = []
    
    for column in df.select_dtypes(include=['float64', 'int64']):
        Q1 = df[column].quantile(0.25)
        Q3 = df[column].quantile(0.75)
        IQR = Q3 - Q1
        
        lower_bound = Q1 - 1.5 * IQR
        upper_bound = Q3 + 1.5 * IQR
        
        outliers = df[(df[column] < lower_bound) | (df[column] > upper_bound)][column]
        
        for outlier in outliers:
            outliers_list.append({'Column': column, 'Outlier Value': outlier})
    
    outliers_df = pd.DataFrame(outliers_list)
    
    return outliers_df

def filter_isolation_forest(df):
    df_numeric = df.select_dtypes(include=['float64', 'int64'])
    df_categorical = df.select_dtypes(include=['object'])

    model = IsolationForest(contamination=0.1, random_state=42)
    model.fit(df_numeric)
    
    # Hacer predicción de outliers (-1 para outliers, 1 para normales)
    df_numeric['outlier'] = model.predict(df_numeric)

    df_concat = pd.concat([df_numeric, df_categorical], axis=1)

    return df_concat


##### Exploración de outliers con método local

outliers_detected = detect_outliers_iqr(df_data) # Detectar outliers en las columnas numéricas
outliers_detected

No se identifican outliers con IQR

#### Detección de outliers con método global

df_outliers = filter_isolation_forest(df_data)

outliers = df_outliers[df_outliers['outlier'] == -1]
outliers

df_outliers['outlier'].value_counts()

df_clean = df_outliers[df_outliers['outlier'] != -1]
df_clean = df_clean.drop('outlier', axis=1)

df_clean

df_clean.to_csv('process_dataset/03_third_result.csv', index=False)
