==> practico2/05_preprocessing_categorical.ipynb <==
import pandas as pd

##### Cambiar valores de columnas a valores numéricos

df_data = pd.read_csv("process_dataset/02_second_result.csv")
df_data

df_data.dtypes

##### Codificación con one-hot
- Trata todas las categorías de manera equitativa, evitando que el modelo aprenda relaciones incorrectas entre ellas.
- Facilita la interpretación

df_data = pd.get_dummies(df_data, columns=['Sex', 'Diet', 'Continent'])
df_data[df_data.select_dtypes(include='bool').columns] = df_data.select_dtypes(include='bool').apply(lambda x: x.astype(int))

df_data.head(5)

df_data.to_csv("process_dataset/05_fifth_result.csv", index=False)
