==> practico2/06_splitting_training_testing.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split

#### Separar dataset en entrenamiento, validación y testeo

df_data = pd.read_csv("process_dataset/05_fifth_result.csv")
df_data.head(5)

#### Separar columna respuesta de los datos

response = df_data['Heart Attack Risk']
data = df_data.drop(columns=['Heart Attack Risk'])

X_train, X_test, y_train, y_test = train_test_split(data, response, test_size=0.1, random_state=42)

X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.2, random_state=42)

#### Guardar datos y columna respuesta

X_train.to_csv("process_dataset/train_data.csv", index=False)
X_val.to_csv("process_dataset/val_data.csv", index=False)
X_test.to_csv("process_dataset/test_data.csv", index=False)

with open("process_dataset/y_train.npy", 'wb') as doc_export:
    np.save(doc_export, y_train)

with open("process_dataset/y_val.npy", 'wb') as doc_export:
    np.save(doc_export, y_val)

with open("process_dataset/y_test.npy", 'wb') as doc_export:
    np.save(doc_export, y_test)
