==> practico2/07_scaler_data.ipynb <==
import pandas as pd
from sklearn.preprocessing import RobustScaler
from joblib import dump

#### Escalar solo el dataset de entrenamiento

df_data = pd.read_csv("process_dataset/train_data.csv")
df_data

#### Separar datos no binarios para escalar
Escalar datos binarios puede alterar su significado y estos ya se encuetran en la escala adecuada (0 y 1)

binary_cols = df_data.columns[df_data.nunique() == 2]
non_binary_cols = df_data.columns.difference(binary_cols)

df_binary = df_data[binary_cols]
df_non_binary = df_data[non_binary_cols]

#### Escalar los datos con Robust Scaler
Se usa Robust Scaler porque los datos tienen aproximadamente un 10% de outliers y es un escalador poco sencible a valores atípicos

def save_scaler(df_data, scaler_instance):
    scaler_instance.fit(X=df_data.values)
    data_scaled = scaler_instance.transform(X=df_data.values)
    df_scaled = pd.DataFrame(data=data_scaled, columns=df_data.columns)
    
    return df_scaled, scaler_instance

robust_scaler_instance = RobustScaler()
df_scaled, scaler_instance = save_scaler(df_non_binary, robust_scaler_instance)

df_scaled

#### Cocatenar datos escalados y no escalados

df_concat = pd.concat([df_binary, df_scaled], axis=1)

df_concat

df_concat.to_csv(f"process_dataset/train_data_scaled_robust.csv", index=False)
dump(scaler_instance, f"results/scaler_robust.joblib")
