==> practico2/08_explore_algorithms.ipynb <==
### Explorar algoritmos
- **Algoritmo de distancias:** k-Nearest Neighbors
- **Algoritmo de reglas:** Árboles de decisión
- **Algoritmo de transformación vectorial:**  SVM
- **Algoritmo de ensamble vía boosting:** Gradient Boosting, AdaBoost
- **Algoritmo de ensamble vía bagging:** Extra-Trees

import pandas as pd
import numpy as np
from joblib import dump

from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import ExtraTreesClassifier

#### Cargar datos desde el práctico anterior

df_data = pd.read_csv("process_dataset/train_data_scaled_robust.csv")
response = np.load("process_dataset/y_train.npy")

#### Instanciar y entrenar los modelos con parámetros por defecto

knn_model = KNeighborsClassifier()
knn_model.fit(X=df_data, y=response)

dt_model = DecisionTreeClassifier()
dt_model.fit(X=df_data, y=response)

svm_model = SVC()
svm_model.fit(X=df_data, y=response)

ada_model = AdaBoostClassifier()
ada_model.fit(X=df_data, y=response)

gbc_model = GradientBoostingClassifier()
gbc_model.fit(X=df_data, y=response)

etc_model = ExtraTreesClassifier()
etc_model.fit(X=df_data, y=response)

#### Guardar los modelos entrenados

dump(knn_model, "results/knn_model.joblib")
dump(dt_model, "results/dt_model.joblib")
dump(svm_model, "results/svm_model.joblib")
dump(ada_model, "results/ada_model.joblib")
dump(gbc_model, "results/gbc_model.joblib")
dump(etc_model, "results/etc_model.joblib")
