==> practico2/13_automatic_optimization.ipynb <==
import pandas as pd
import numpy as np

from tpot import TPOTClassifier

#### Método de optimización automática

X_train = pd.read_csv("process_dataset/train_data_scaled_robust.csv")
y_train = np.load("process_dataset/y_train.npy")

X_test = pd.read_csv("process_dataset/test_data.csv")
y_test = np.load("process_dataset/y_test.npy")

X_val = pd.read_csv("process_dataset/val_data.csv")
y_val = np.load("process_dataset/y_val.npy")

#### Optimización indicando el algoritmo

custom_config = {
    'sklearn.tree.DecisionTreeClassifier': {
        'criterion': ['gini', 'entropy'],
        'max_depth': range(1, 20),
        'min_samples_split': range(2, 20),
        'min_samples_leaf': range(1, 20)
    }
}

tpot = TPOTClassifier(
    config_dict=custom_config,
    generations=20,           
    population_size=20,        
    verbosity=2,                
    random_state=42              
)

tpot.fit(X_train, y_train)

#### Optimización sin restricciones

tpot = TPOTClassifier(
    generations=50,              # Número de generaciones (más generaciones -> mejor optimización)
    population_size=50,         # Tamaño de la población (mayor población -> mejor exploración)
    verbosity=2,               
    random_state=42            
)

tpot.fit(X_train, y_train)
